Fixed location and improved webhook logging

+++ Modules/CIPPCore/Public/Invoke-CIPPGraphWebhookRenewal.ps1
@@ -28,10 +28,13 @@ function Invoke-CippGraphWebhookRenewal {
                 $Resource = "$($UpdateSub.Resource)"
                 $EventType = "$($UpdateSub.EventType)"
 
-                Remove-AzDataTableEntity @WebhookTable -Entity $UpdateSub
                 Write-LogMessage -user 'CIPP' -API 'Renew_Graph_Subscriptions' -message "Recreating: $($UpdateSub.SubscriptionID) as renewal failed." -Sev "Info" -tenant $TenantFilter
+                $CreateResult = New-CIPPGraphSubscription -TenantFilter $TenantFilter -TypeofSubscription $TypeofSubscription -BaseURL $BaseURL -Resource $Resource -EventType $EventType -ExecutingUser 'GraphSubscriptionRenewal'
+
+                if ($CreateResult -match 'Created Webhook subscription for') {
+                    Remove-AzDataTableEntity @WebhookTable -Entity $UpdateSub
+                }
                 
-                New-CIPPGraphSubscription -TenantFilter $TenantFilter -TypeofSubscription $TypeofSubscription -BaseURL $BaseURL -Resource $Resource -EventType $EventType -ExecutingUser 'GraphSubscriptionRenewal'
             }
